changes from jari

+++ app/src/main/java/com/teamx/equiz/ui/fragments/topup/data/TopUpModelData.kt
@@ -5,6 +5,6 @@ import androidx.annotation.Keep
 @Keep
 data class TopUpModelData(
     val checkout: String,
-    val `data`: Int,
+    val `data`: Double,
     val message: String
 )
